add docker file and rebrand
