+++ pom.xml
@@ -12,6 +12,7 @@
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
         <version>2.6.3</version>
+        <relativePath></relativePath>
     </parent>
 
     <!-- project properties -->
