+++ setup.ini
@@ -1,5 +1,5 @@
 [setup]
-serverName = antony
+serverName = Ward
 theme = light
 port = 5000
