--- src/main/java/org/bsoftware/ward/exceptions/ApplicationNotConfiguredException.java
@@ -1,18 +1,0 @@
-package org.bsoftware.ward.exceptions;
-
-/**
- * ApplicationNotSetUpException indicates that user tried to access api, without setting up the application
- *
- * @author Rudolf Barbu
- * @version 1.0.0
- */
-public final class ApplicationNotConfiguredException extends Exception
-{
-    /**
-     * Call super class with exception message
-     */
-    public ApplicationNotConfiguredException()
-    {
-        super("Set up application first");
-    }
-}
